inserted content of prism.c into geom.c, made updates to enable mpb and meep builds
