+++ utils/ctlgeom.h
@@ -152,19 +152,6 @@ GEOMETRIC_OBJECT make_ellipsoid(MATERIAL_TYPE material, vector3 center,
 GEOMETRIC_OBJECT make_prism(MATERIAL_TYPE material,
 			    const vector3 *vertices, int num_vertices,
 			    double height, vector3 axis);
-
-vector3 prism_coordinate_p2c(prism *prsm, vector3 vp);
-vector3 prism_coordinate_c2p(prism *prsm, vector3 vc);
-vector3 prism_vector_p2c(prism *prsm, vector3 vp);
-vector3 prism_vector_c2p(prism *prsm, vector3 vc);
-void get_prism_bounding_box(prism *prsm, geom_box *box);
-double intersect_line_segment_with_prism(prism *prsm, vector3 p, vector3 d, double a, double b);
-boolean point_in_polygon(double px, double py, vector3 *vertices, int num_vertices);
-boolean point_in_prism(prism *prsm, vector3 p);
-vector3 normal_to_prism(prism *prsm, vector3 p);
-void display_prism_info(int indentby, prism *prsm);
-int vector3_nearly_equal(vector3 v1, vector3 v2);
-
 /**************************************************************************/
 
 #ifdef __cplusplus
